Move changes from public interface to extension method
