+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/IProject.cs
@@ -21,8 +21,6 @@ namespace Microsoft.DotNet.UpgradeAssistant
 
         IEnumerable<IProject> ProjectReferences { get; }
 
-        public IEnumerable<string> AllProjectReferences { get; }
-
         IEnumerable<Reference> FrameworkReferences { get; }
 
         INuGetReferences NuGetReferences { get; }
